requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
numpy
pandas
scikit-learn

# Parsers
pdfplumber
python-docx
pymupdf
pdfminer.six

# Optional
spacy
rapidfuzz

# FAISS backend (optional; Windows users may skip and use in-mem)
faiss-cpu
pydantic
typer
